requirements.txt
langgraph>=0.0.40
google-genai>=0.3.0
pydantic>=2.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
